don't set 'acl' flag globally
Enabling this flag does not magically enable the use of access control
lists.  For this additional userspace tools have to be installed and
support has to enabled in the kernel.  Users capable of doing this and
especially just considering that they might want to enable this on
their machines can be expected to check for the existing useflag also.