sys-kernel/openvz-rhel6-stable: Support building with GCC 4.6.2
Fix build failures by cherry-picking a few patches from newer kernels.
Note that rhel6-openvz-sources-fix-build-without-xen.patch is strictly
not necessary, but it is included for completeness should anyone wish to
build a custom kernel without Xen support.

We do a revision bump and drop keywords so that the ebuilds are
available for testing, but are not yet available to the general public.
Keywords will be added after testing is finished.

Signed-off-by: Richard Yao <ryao@funtoo.org>