Masking electron-1.4.15 until FL-3545 gets resolved
Build failures caused by missing cups patching and
the incompatibility of system jinja-2.9 with chromium 53
used to build electron cause build failures. The new
ebuild is a work in progress and the issue is being actively
investigated. Currently the new electron version is not being
used by any package in the portage tree, however, it's being
masked as a precautionary measure.