Fix pouetchess compilation failures with gcc6
Patched pouetchess to avoid errors caused by
stricter behavior of gcc6. The changes are minor
and shouldn't cause any breakage.

Current users are unaffected. The current revision
of the package is therefore kept.

Tested to be working with gcc-5.3.0 and gcc-6.3.0

Relevant issue:
https://bugs.funtoo.org/browse/FL-3455

Issue should be reported to Gentoo and the package
should be unforked when Gentoo applies the patches.

Commited by: <hencb @ protonmail .ch>