Fork virt-manager to fix broken dep until shard update
virt-manager is not sharded but requires
libvirt-glib (sharded)  as a runtime dependency.
The Gentoo update to virt-manager is broken
because there is no libvirt-glib-1.0.0 in Funtoo:
https://github.com/gentoo/gentoo/commit/6db7f05662e9405b40759d00b97433def6048779
A shard update to include the update to libvirt-glib-1.0.0
would be unreasonable.

To fix this, I'm forking virt-manager to funtoo-overlay
and reverting to the old behavior of the ebuild.
Forking libvirt-glib instead of virt-manager
could have unforseen consequences and therefore
I'm sticking to the conservative option.

When the gnome shard updates to this commit:
http://git.funtoo.org/gentoo-gnome-shard/commit/?id=3a83d7e6437120f0fcf180d73b5ca939ddf0e76c
the virt-manager ebuild needs to be unforked
to avoid breakage.

Added a note about this to my funtoo-reports repo,
so that we don't forget to unfork this commit:
https://github.com/antematherian/funtoo-reports/blob/master/unfork-list/march2017.txt

Commited by: Bruno Henc <hencb @ protonmail.ch>