Add ebuilds for dia and mpaste from bhenc
mpaste is a pastebin client that interfaces with the pastebin
service hosted by bhenc at paste.mheaven.ch or any equivalent
pasteme based pastebin server. It's used for app-text/dia.

app-text/dia is a diagnostic script written by bhenc to ease
Funtoo support. It runs a curated set of commands and uploads
their output to chat.mheaven.ch using the mpaste client. The
reason I'm not using wgetpaste is that I'm not comfortable
a) sharing the diagnostic data to pastebin services hosted by
third parties b) the script generates a lot of links and therefore
it seems unappropriate to spam third-party pastebin services.
It could even hit some kind of rate-limiting that could be
problematic.

Both dia and mpaste have been tested to work properly with Funtoo.
Dia was tested on a running system, on a Funtoo container, and in
a chroot. Commands do fail on the container, however, that doesn't
prevent the script from collecting data.

I'm open for suggestions on imporving the script. I find that both
mpaste and dia work well and that they are ready forbasic usage.