FL-3871: restore iwlxxxx-ucode ebuilds
Gentoo seems to have treecleaned those
ebuilds because the intellinuxwireless
project moved to the linux wirelsss
project, and all SRC_URI links died.

I updated all the SRC_URIs and made sure
the ebuilds worked and use the up-to-date
firmwares. All the firmwares are also mirrored
for good measure on the ceresia mirrors.

All the ebuilds should install correctly.
The iwl-4965-ucode firmware works, currently
running on my R61 thinkpad, and the firmware
is equvilanet to the one in linux-firmware.

Users should have the option to only install
the firmware they need without having to download
the huge linux firmware source tarball. This commit
restores this option for the intel wireless cards.