Update ports-2012 portage to ease kit migration
The portage update to -r4 fixes two nasty issues, and
adds two portage features that are handy.

The obvious is patching around setup.py to stop including
the gentoo repos.conf in /var/lib/portage/repos.conf

The less obvious is fixing the default git behavior to
pull from remote only the current branch while doing
a shallow clone of a git repository.

Simply put:
a) git clone https://github.com/funtoo/core-kit
will pull all branches.

b) git clone --depth=1 https://github.com/funtoo/core-kit
will only pull whatever is set to be the main branch.
So doing git checkout 1.0-prime on b) won't work.

however c) git clone --depth=1 --no-single-branch https://
github.com/funtoo/core-kit
will pull down the current tips of all branches so
git checkout 1.0-prime and git checkout master will work
nicely.

This is what the sync-branch-sync-enforcing patch includes
by adding --no-single-branch to git_cmd_opts on line 57
of the patch.

The other two thing sync-branch-sync-enforcing does is add
two new options for repos.conf entries that use git:

sync-branch as the name implies will sync the specified
branch if auto-sync = yes is set. This is useful since
correctly cloning a branch initially avoids messing around
with git branches.

auto-sync-enforcing = yes enforces the values set in repos.conf
entries. It makes sure that the git branches are set exactly
in the way repos.conf entries specify, and portage will take
drastic action to ensure that those values are correct. This is
important since the portage API allows loading the repos.conf
entries for lookup by external programs. If the repos.conf entries
do not match the git set values, things get nasty pretty quickly
because we have a SPOT (Single Point Of Truth) violation.

So if sync-branch doesn't match git branch -v, portage switches to
the correct branches and issues a git pull command to update the
branch. By default emerge --sync only updates the current branch
(it really inherits this behavior from git pull)

If the sync-uri doesn't match, portage NUKES THE REPOSITORY, and
redownloads it. This will automatically set the correct branch.

This behavior is enforced every time emerge --sync is done. So
to change a branch, or sync-uri, one simply has to change the
values in the appropriate repos.conf entry, and run emerge --sync.
This is awfully convenient if you want to bail yourself out of a
incorrect-branch-being-set-on-multiple-repositories-nightmare.
This likes to happen if you Ctrl+C out of a script that does branch
switching on multiple repositories, and will happen if that kind
of functionality is added to portage (branch switching is cool for
automatic resolving of missing dependencies, which will be needed
to address the stale dependency problem with some ebuilds).

If auto-sync-enforcing = yes and auto-sync = yes is not set, it's
the responsibility of the developer to interface with git directly.
Same goes for the user: it's his responsibility to keep track of
the git branches since portage inherently isn't too git aware, except
if the auto-sync switch is turned on. As far as portage is concerned,
by default, a repository is a collection of files.