I have changed how merge-all-kits parses args (needs testing) by using argparse and improving the calling convention. Use as follows:
--db to enable "update fastpull db" (defaults to off.) --nopush to
disable pushing (defaults to on.) And if nothing is specified as
parameter on command-line, then all releases will be generated, but
if a single parameter is specified like "1.3-release", then only this
release will be generated, saving time. Now that our releases are
"uncoupled" from each other, it's now possible to only update one
release at a time.